+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - ISY-451: Die Dokumentation der Bibliothek ist auf GitHub Pages verfügbar
 - ISY-469: Es wurden einige Probleme mit der Barrierefreiheit entfernt
 ## Demo-Anwendung
+- ISY-378: Hinzufügen von Validatoren auf der Objekt-Anzeigen-Seite
 - ISY-360: Verhalten nach Navigation auf eine nicht existierende Seite
 - ISY-516: Entfernen von automatischen Ausblenden von Toast-Notifications